Fp-kontrakter bumpes til 6.2.1 hvor avhengigheter til fpwsproxy har package rename
