+++ domenetjenester/okonomistotte/src/main/java/no/nav/foreldrepenger/økonomistøtte/simulering/klient/OppdragsKontrollDtoMapper.java
@@ -13,20 +13,20 @@ import no.nav.foreldrepenger.behandlingslager.økonomioppdrag.Oppdragslinje150;
 import no.nav.foreldrepenger.behandlingslager.økonomioppdrag.Refusjonsinfo156;
 import no.nav.foreldrepenger.behandlingslager.økonomioppdrag.Sats;
 import no.nav.foreldrepenger.behandlingslager.økonomioppdrag.Utbetalingsgrad;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeEndring;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeEndringLinje;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeFagområde;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeKlassifik;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeStatusLinje;
-import no.nav.foreldrepenger.kontrakter.simulering.request.LukketPeriode;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Ompostering116Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Oppdrag110Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.OppdragskontrollDto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Oppdragslinje150Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Refusjonsinfo156Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.SatsDto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.TypeSats;
-import no.nav.foreldrepenger.kontrakter.simulering.request.UtbetalingsgradDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeEndring;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeEndringLinje;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeFagområde;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeKlassifik;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeStatusLinje;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.LukketPeriode;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Ompostering116Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Oppdrag110Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.OppdragskontrollDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Oppdragslinje150Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Refusjonsinfo156Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.SatsDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.TypeSats;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.UtbetalingsgradDto;
 
 public class OppdragsKontrollDtoMapper {
