+++ pom.xml
@@ -39,7 +39,7 @@
         <prosesstask.version>3.1.14</prosesstask.version>
         <fp-jms.version>2.2.0</fp-jms.version>
 
-        <fp-kontrakter.version>6.1.31</fp-kontrakter.version>
+        <fp-kontrakter.version>6.2.1</fp-kontrakter.version>
         <fp-tidsserie.version>2.6.5</fp-tidsserie.version>
         <fp-nare.version>2.4.0</fp-nare.version>
         <abakus-kontrakt.version>1.3.37</abakus-kontrakt.version>
